Revert "New landing page (#110)"
This reverts commit c6b287ef03cd385ea78280ee7fa9da7fea5b9bcd.
