+++ tailwind.config.js
@@ -7,7 +7,6 @@ module.exports = {
     extend: {
       fontFamily: {
         sans: ["Jeko", "sans-serif"],
-        roboto: ["Roboto"]
       },
       textUnderlineOffset: {
         16: "16px",
